+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/lib/arguments/PlayerByNameArgument.java
@@ -1,6 +1,5 @@
 package com.nisovin.shopkeepers.commands.lib.arguments;
 
-import java.util.Optional;
 import java.util.stream.Stream;
 
 import org.bukkit.entity.Player;
@@ -9,7 +8,9 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 
 import com.nisovin.shopkeepers.commands.lib.CommandInput;
 import com.nisovin.shopkeepers.commands.lib.argument.ArgumentParseException;
+import com.nisovin.shopkeepers.commands.lib.argument.ambiguity.AmbiguousInputHandler;
 import com.nisovin.shopkeepers.commands.lib.argument.filter.ArgumentFilter;
+import com.nisovin.shopkeepers.commands.lib.argument.filter.ArgumentRejectedException;
 import com.nisovin.shopkeepers.commands.lib.context.CommandContextView;
 import com.nisovin.shopkeepers.commands.lib.util.PlayerArgumentUtils.PlayerNameMatcher;
 import com.nisovin.shopkeepers.lang.Messages;
@@ -68,17 +69,67 @@ public class PlayerByNameArgument extends ObjectByIdArgument<String, Player> {
 		return Messages.commandPlayerArgumentInvalid;
 	}
 
+	/**
+	 * Gets the {@link AmbiguousInputHandler} for players matched by name.
+	 * <p>
+	 * When overriding this method, consider applying the {@link #getDefaultErrorMsgArgs() common
+	 * message arguments} to the error message returned by the {@link AmbiguousInputHandler} (if
+	 * any).
+	 * 
+	 * @param argumentInput
+	 *            the argument input
+	 * @param matchedPlayers
+	 *            the matched players
+	 * @return the ambiguous player name handler, not <code>null</code>
+	 */
+	protected AmbiguousInputHandler<Player> getAmbiguousPlayerNameHandler(
+			String argumentInput,
+			Iterable<? extends Player> matchedPlayers
+	) {
+		var ambiguousPlayerNameHandler = new AmbiguousPlayerNameHandler<Player>(
+				argumentInput,
+				matchedPlayers
+		);
+		if (ambiguousPlayerNameHandler.isInputAmbiguous()) {
+			// Apply common message arguments:
+			Text errorMsg = ambiguousPlayerNameHandler.getErrorMsg();
+			assert errorMsg != null;
+			errorMsg.setPlaceholderArguments(this.getDefaultErrorMsgArgs());
+			errorMsg.setPlaceholderArguments("argument", argumentInput);
+		}
+		return ambiguousPlayerNameHandler;
+	}
+
+	/**
+	 * The default implementation of getting a {@link Player} by name.
+	 * 
+	 * @param nameInput
+	 *            the name input
+	 * @return the matched player, or <code>null</code> if no match is found
+	 * @throws ArgumentRejectedException
+	 *             if the name is ambiguous
+	 */
+	public final @Nullable Player getDefaultPlayerByName(String nameInput)
+			throws ArgumentRejectedException {
+		// The name input can be either player name or display name:
+		Stream<Player> players = PlayerNameMatcher.EXACT.match(nameInput);
+		var ambiguousPlayerNameHandler = this.getAmbiguousPlayerNameHandler(nameInput, players::iterator);
+		if (ambiguousPlayerNameHandler.isInputAmbiguous()) {
+			Text errorMsg = ambiguousPlayerNameHandler.getErrorMsg();
+			assert errorMsg != null;
+			throw new ArgumentRejectedException(this, errorMsg);
+		} else {
+			return ambiguousPlayerNameHandler.getFirstMatch();
+		}
+	}
+
 	@Override
 	protected @Nullable Player getObject(
 			CommandInput input,
 			CommandContextView context,
 			String nameInput
 	) throws ArgumentParseException {
-		// Name input may be both player name or display name:
-		Stream<Player> players = PlayerNameMatcher.EXACT.match(nameInput);
-		Optional<Player> player = players.findFirst();
-		return player.orElse(null);
-		// TODO deal with ambiguities
+		return this.getDefaultPlayerByName(nameInput);
 	}
 
 	@Override